Move KotlinEnvironment specific part to internal package
